Pod-1 worker pool size tweak
The change in https://github.com/travis-ci/packer-templates/pull/630
introduces a dedicated macpro to run packer builds for xcode build
images for the mac infra. We decided to provision this macpro from the
existing `Macpro_Pod_1` shared compute cluster.  As a result, worker
pool sizes need to change so that the remaining macpros do not get
overencumbered during peak demand.