Deploy new VM images for MacStadium Kubernetes nodes
We don't want to replace the master because it's not under disk pressure
and we would lose the existing configuration. So now nodes and master
can have different VM templates.

Nodes now use a CentOS 7 template with Kubernetes already installed
during the Packer build, so the provisioning is substantially simpler.