Upgrade GKE clusters to 1.13
The upgrade was to implement patches for NFLX-2019-001. [1] I used this
opportunity to also upgrade the clusters to a new Kubernetes version.

Upgrading to 1.12 and then to 1.13, this happened in a few steps:
1. Upgrade master to 1.12 with Terraform
2. Upgrade nodes with gcloud container clusters upgrade <cluster>
3. Upgrade master to 1.13 with Terraform
4. Upgrade nodes with gcloud container clusters upgrade <cluster>

The master upgrade happens without impacting the nodes, they stay up
during the upgrade. The node upgrades will reschedule the node to a new
node for a while and schedule it back to the node where it came from,
this limits the downtime.

However, once we start running workers, upgrading the nodes should be
done with more care. Therefore, I have disabled auto-upgrade on the
nodes. Read more about this in worker-operator[2].

I've also changed the parameter 'zone' to 'location' to resolve the
terraform deprecation warning and renamed the variable to region as we
should use this to set a region in the future instead of a zone, imo.

[1] https://cloud.google.com/kubernetes-engine/docs/security-bulletins#june-24-2019-2
[2] https://github.com/travis-ci/worker-operator