Increase resources for apt/build-cache
A. 200GB disk * 0.8 = 160GB space for caching, this should give us a
better performance disk and higher throughput [1]

B. 360m or 6 hours of cache time, this should keep cache around a bit
longer.

C. 2vCPU / 8GB RAM -> This is half the CPU it has now, but we will double
the instances (2 more on prod-1), so this should be good.

Change A. and B. should increase the HIT/MISS ratio on Nginx, lowering the
latency and decreasing network traffice over time.

[1] https://cloud.google.com/compute/docs/disks/performance